event/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.5.0
asn1crypto==1.5.1
attrs==23.1.0
boto3==1.21.35
botocore==1.24.35
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
cssselect2==0.7.0
dj-database-url==0.5.0
Django==4.0.3
django-phone-field==1.8.1
django-storages==1.12.3
django-xhtml2pdf==0.0.4
fonttools==4.46.0
freetype-py==2.3.0
frozenlist==1.4.0
gunicorn==20.1.0
html5lib==1.1
idna==3.6
jmespath==1.0.0
lxml==4.9.3
multidict==6.0.4
nodeenv==1.8.0
oscrypto==1.3.0
Pillow==10.1.0
pycairo==1.25.1
pycparser==2.21
pydyf==0.8.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==3.17.3
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
PyYAML==6.0.1
reportlab==4.0.8
requests==2.31.0
rlPyCairo==0.3.0
s3transfer==0.5.2
segno==1.4.1
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
svglib==1.5.1
twilio==8.10.3
typing_extensions==4.9.0
tzdata==2022.1
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.13